Abstracted the Scanning Code into an extension
- for readability
